requirements-remote.txt
# Core Dependencies for BEV on Remote Nodes
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Async & Networking
asyncio==3.4.3
aiohttp==3.9.1
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0

# Database Connectors
asyncpg==0.29.0
redis==5.0.1
neo4j==5.15.0
psycopg2-binary==2.9.9
pymongo==4.6.0
influxdb-client==1.38.0

# Message Queues
pika==1.3.2
kafka-python==2.0.2
celery==5.3.4

# ML/AI (for GPU nodes)
torch==2.1.0
transformers==4.36.0
sentence-transformers==2.2.2
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.16.0
scrapy==2.11.0

# Security & Cryptography
cryptography==41.0.7
pycryptodome==3.19.0
hvac==1.2.1

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# API Frameworks
fastapi==0.105.0
uvicorn[standard]==0.24.0
flask==3.0.0

# Vector Databases
qdrant-client==1.7.0
weaviate-client==4.4.0

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
pyyaml==6.0.1
python-multipart==0.0.6
python-json-logger==2.0.7
colorama==0.4.6
tqdm==4.66.1
click==8.1.7
